pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=3.0.0",
    "django>=5.2.1",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "flake8>=7.2.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "whitenoise[brotli]>=6.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
